split local settings gui into smaller parts
