+++ pype/tools/settings/local_settings/window.py
@@ -1,27 +1,26 @@
-import os
-import sys
-import traceback
-import platform
 import logging
 from Qt import QtWidgets, QtCore, QtGui
-from pype.tools.settings.settings.widgets.widgets import (
-    ExpandingWidget,
-    SpacerWidget,
-    ProjectListWidget
-)
+
 from ..settings import style
-from pype.tools.settings.settings.widgets.lib import CHILD_OFFSET
-from pype.settings.constants import PROJECT_ANATOMY_KEY
+
 from pype.settings.lib import (
     get_local_settings,
     save_local_settings
 )
 from pype.api import (
     SystemSettings,
-    ProjectSettings,
-    change_pype_mongo_url
+    ProjectSettings
 )
-from pymongo.errors import ServerSelectionTimeoutError
+
+from .widgets import (
+    Separator,
+    SpacerWidget,
+    ExpandingWidget,
+    CHILD_OFFSET
+)
+from .mongo_widget import PypeMongoWidget
+from .apps_widget import LocalApplicationsWidgets
+from .projects_widget import ProjectSettingsWidget
 
 log = logging.getLogger(__name__)
 
@@ -30,94 +29,6 @@ LOCAL_GENERAL_KEY = "general"
 LOCAL_PROJECTS_KEY = "projects"
 LOCAL_APPS_KEY = "applications"
 LOCAL_ROOTS_KEY = "roots"
-
-
-class Separator(QtWidgets.QFrame):
-    def __init__(self, height=None, parent=None):
-        super(Separator, self).__init__(parent)
-        if height is None:
-            height = 2
-
-        splitter_item = QtWidgets.QWidget(self)
-        splitter_item.setStyleSheet("background-color: #21252B;")
-        splitter_item.setMinimumHeight(height)
-        splitter_item.setMaximumHeight(height)
-
-        layout = QtWidgets.QHBoxLayout(self)
-        layout.setContentsMargins(5, 5, 5, 5)
-        layout.addWidget(splitter_item)
-
-
-class PypeMongoWidget(QtWidgets.QWidget):
-    def __init__(self, parent):
-        super(PypeMongoWidget, self).__init__(parent)
-
-        # Warning label
-        warning_label = QtWidgets.QLabel((
-            "WARNING: Requires restart. Change of Pype Mongo requires to"
-            " restart of all running Pype processes and process using Pype"
-            " (Including this)."
-            "\n- all changes in different categories won't be saved."
-        ), self)
-        warning_label.setStyleSheet("font-weight: bold;")
-
-        # Label
-        mongo_url_label = QtWidgets.QLabel("Pype Mongo URL", self)
-
-        # Input
-        mongo_url_input = QtWidgets.QLineEdit(self)
-        mongo_url_input.setPlaceholderText("< Pype Mongo URL >")
-        mongo_url_input.setText(os.environ["PYPE_MONGO"])
-
-        # Confirm button
-        mongo_url_change_btn = QtWidgets.QPushButton("Confirm Change", self)
-
-        layout = QtWidgets.QGridLayout(self)
-        layout.setContentsMargins(0, 0, 0, 0)
-        layout.addWidget(warning_label, 0, 0, 1, 3)
-        layout.addWidget(mongo_url_label, 1, 0)
-        layout.addWidget(mongo_url_input, 1, 1)
-        layout.addWidget(mongo_url_change_btn, 1, 2)
-
-        mongo_url_change_btn.clicked.connect(self._on_confirm_click)
-
-        self.mongo_url_input = mongo_url_input
-
-    def _on_confirm_click(self):
-        value = self.mongo_url_input.text()
-
-        dialog = QtWidgets.QMessageBox(self)
-
-        title = "Pype mongo changed"
-        message = (
-            "Pype mongo url was successfully changed. Restart Pype please."
-        )
-        details = None
-
-        try:
-            change_pype_mongo_url(value)
-        except Exception as exc:
-            if isinstance(exc, ServerSelectionTimeoutError):
-                error_message = (
-                    "Connection timeout passed."
-                    " Probably can't connect to the Mongo server."
-                )
-            else:
-                error_message = str(exc)
-
-            title = "Pype mongo change failed"
-            # TODO catch exception message more gracefully
-            message = (
-                "Pype mongo change was not successful."
-                " Full traceback can be found in details section.\n\n"
-                "Error message:\n{}"
-            ).format(error_message)
-            details = "\n".join(traceback.format_exception(*sys.exc_info()))
-        dialog.setWindowTitle(title)
-        dialog.setText(message)
-        if details:
-            dialog.setDetailedText(details)
-        dialog.exec_()
 
 
 class LocalGeneralWidgets(QtWidgets.QWidget):
@@ -148,407 +59,6 @@ class LocalGeneralWidgets(QtWidgets.QWidget):
             output["site_name"] = local_site_name
         # Do not return output yet since we don't have mechanism to save or
         #   load these data through api calls
-        return output
-
-
-class AppVariantWidget(QtWidgets.QWidget):
-    exec_placeholder = "< Specific path for this machine >"
-
-    def __init__(self, group_label, variant_entity, parent):
-        super(AppVariantWidget, self).__init__(parent)
-
-        self.executable_input_widget = None
-
-        label = " ".join([group_label, variant_entity.label])
-
-        expading_widget = ExpandingWidget(label, self)
-        content_widget = QtWidgets.QWidget(expading_widget)
-        content_layout = QtWidgets.QVBoxLayout(content_widget)
-        content_layout.setContentsMargins(CHILD_OFFSET, 5, 0, 0)
-
-        expading_widget.set_content_widget(content_widget)
-
-        # Add expanding widget to main layout
-        layout = QtWidgets.QHBoxLayout(self)
-        layout.setContentsMargins(0, 0, 0, 0)
-        layout.addWidget(expading_widget)
-
-        # TODO For celaction - not sure what is "Celaction publish" for
-        if not variant_entity["executables"].multiplatform:
-            warn_label = QtWidgets.QLabel(
-                "Application without multiplatform paths"
-            )
-            content_layout.addWidget(warn_label)
-            return
-
-        executable_input_widget = QtWidgets.QLineEdit(content_widget)
-        executable_input_widget.setPlaceholderText(self.exec_placeholder)
-        content_layout.addWidget(executable_input_widget)
-
-        self.executable_input_widget = executable_input_widget
-
-        studio_executables = (
-            variant_entity["executables"][platform.system().lower()]
-        )
-        if len(studio_executables) < 1:
-            return
-
-        content_layout.addWidget(Separator(parent=self))
-        content_layout.addWidget(
-            QtWidgets.QLabel("Studio paths:", self)
-        )
-
-        for item in studio_executables:
-            path_widget = QtWidgets.QLineEdit(content_widget)
-            path_widget.setText(item.value[0])
-            path_widget.setEnabled(False)
-            content_layout.addWidget(path_widget)
-
-    def set_value(self, value):
-        if not self.executable_input_widget:
-            return
-
-        if not value:
-            value = {}
-        elif not isinstance(value, dict):
-            print("Got invalid value type {}. Expected {}".format(
-                type(value), dict
-            ))
-            value = {}
-
-        executable_path = value.get("executable")
-        if not executable_path:
-            executable_path = ""
-        elif isinstance(executable_path, list):
-            print("Got list in executable path so using first item as value")
-            executable_path = executable_path[0]
-
-        if not isinstance(executable_path, str):
-            executable_path = ""
-            print((
-                "Got invalid value type of app executable {}. Expected {}"
-            ).format(type(value), str))
-
-        self.executable_input_widget.setText(executable_path)
-
-    def settings_value(self):
-        if not self.executable_input_widget:
-            return None
-        value = self.executable_input_widget.text()
-        if not value:
-            return None
-        return {"executable": value}
-
-
-class AppGroupWidget(QtWidgets.QWidget):
-    def __init__(self, group_entity, parent):
-        super(AppGroupWidget, self).__init__(parent)
-
-        valid_variants = {}
-        for key, entity in group_entity["variants"].items():
-            if entity["enabled"].value:
-                valid_variants[key] = entity
-
-        group_label = group_entity.label
-        expading_widget = ExpandingWidget(group_label, self)
-        content_widget = QtWidgets.QWidget(expading_widget)
-        content_layout = QtWidgets.QVBoxLayout(content_widget)
-        content_layout.setContentsMargins(CHILD_OFFSET, 5, 0, 0)
-
-        widgets_by_variant_name = {}
-        for variant_name, variant_entity in valid_variants.items():
-            variant_widget = AppVariantWidget(
-                group_label, variant_entity, content_widget
-            )
-            widgets_by_variant_name[variant_name] = variant_widget
-            content_layout.addWidget(variant_widget)
-
-        expading_widget.set_content_widget(content_widget)
-
-        layout = QtWidgets.QHBoxLayout(self)
-        layout.setContentsMargins(0, 0, 0, 0)
-        layout.addWidget(expading_widget)
-
-        self.widgets_by_variant_name = widgets_by_variant_name
-
-    def set_value(self, value):
-        if not value:
-            value = {}
-
-        for variant_name, widget in self.widgets_by_variant_name.items():
-            widget.set_value(value.get(variant_name))
-
-    def settings_value(self):
-        output = {}
-        for variant_name, widget in self.widgets_by_variant_name.items():
-            value = widget.settings_value()
-            if value:
-                output[variant_name] = value
-
-        if not output:
-            return None
-        return output
-
-
-class LocalApplicationsWidgets(QtWidgets.QWidget):
-    def __init__(self, system_settings_entity, parent):
-        super(LocalApplicationsWidgets, self).__init__(parent)
-
-        widgets_by_group_name = {}
-
-        layout = QtWidgets.QVBoxLayout(self)
-        layout.setContentsMargins(0, 0, 0, 0)
-
-        for key, entity in system_settings_entity["applications"].items():
-            # Filter not enabled app groups
-            if not entity["enabled"].value:
-                continue
-
-            # Check if has enabled any variant
-            enabled_variant = False
-            for variant_entity in entity["variants"].values():
-                if variant_entity["enabled"].value:
-                    enabled_variant = True
-                    break
-
-            if not enabled_variant:
-                continue
-
-            # Create App group specific widget and store it by the key
-            group_widget = AppGroupWidget(entity, self)
-            widgets_by_group_name[key] = group_widget
-            layout.addWidget(group_widget)
-
-        self.widgets_by_group_name = widgets_by_group_name
-
-    def set_value(self, value):
-        if not value:
-            value = {}
-
-        for group_name, widget in self.widgets_by_group_name.items():
-            widget.set_value(value.get(group_name))
-
-    def settings_value(self):
-        output = {}
-        for group_name, widget in self.widgets_by_group_name.items():
-            value = widget.settings_value()
-            if value:
-                output[group_name] = value
-        if not output:
-            return None
-        return output
-
-
-class RootsWidget(QtWidgets.QWidget):
-    value_changed = QtCore.Signal()
-
-    def __init__(self, project_settings, parent):
-        self._parent_widget = parent
-        super(RootsWidget, self).__init__(parent)
-
-        self.project_settings = project_settings
-        self.widgts_by_root_name = {}
-
-        main_layout = QtWidgets.QVBoxLayout(self)
-
-        self.content_layout = main_layout
-
-    def _on_root_value_change(self):
-        self.value_changed.emit()
-
-    def refresh(self):
-        while self.content_layout.count():
-            item = self.content_layout.itemAt(0)
-            item.widget().hide()
-            self.content_layout.removeItem(item)
-
-        self.widgts_by_root_name.clear()
-
-        default_root_values = self.local_default_project_values() or {}
-
-        roots_entity = (
-            self.project_settings[PROJECT_ANATOMY_KEY][LOCAL_ROOTS_KEY]
-        )
-        is_in_default = self.project_settings.project_name is None
-        for root_name, path_entity in roots_entity.items():
-            platform_entity = path_entity[platform.system().lower()]
-            root_widget = QtWidgets.QWidget(self)
-
-            key_label = QtWidgets.QLabel(root_name, root_widget)
-
-            root_input_widget = QtWidgets.QWidget(root_widget)
-            root_input_layout = QtWidgets.QVBoxLayout(root_input_widget)
-
-            value_input = QtWidgets.QLineEdit(root_input_widget)
-            placeholder = None
-            if not is_in_default:
-                placeholder = default_root_values.get(root_name)
-                if placeholder:
-                    placeholder = "< {} >".format(placeholder)
-
-            if not placeholder:
-                placeholder = platform_entity.value
-            value_input.setPlaceholderText(placeholder)
-            value_input.textChanged.connect(self._on_root_value_change)
-
-            root_input_layout.addWidget(value_input)
-
-            root_layout = QtWidgets.QHBoxLayout(root_widget)
-            root_layout.addWidget(key_label)
-            root_layout.addWidget(root_input_widget)
-
-            self.content_layout.addWidget(root_widget)
-            self.widgts_by_root_name[root_name] = value_input
-
-        self.content_layout.addWidget(SpacerWidget(self), 1)
-
-    def local_default_project_values(self):
-        default_project = self._parent_widget.per_project_settings.get(None)
-        if default_project:
-            return default_project.get(LOCAL_ROOTS_KEY)
-        return None
-
-    def set_value(self, value):
-        if not value:
-            value = {}
-
-        for root_name, widget in self.widgts_by_root_name.items():
-            root_value = value.get(root_name) or ""
-            widget.setText(root_value)
-
-    def settings_value(self):
-        output = {}
-        for root_name, widget in self.widgts_by_root_name.items():
-            value = widget.text()
-            if value:
-                output[root_name] = value
-        if not output:
-            return None
-        return output
-
-
-class ProjectSpecificWidget(QtWidgets.QWidget):
-    value_changed = QtCore.Signal()
-
-    def __init__(self, project_settings, parent):
-        self._parent_widget = parent
-        super(ProjectSpecificWidget, self).__init__(parent)
-
-        self.project_settings = project_settings
-        self.widgts_by_root_name = {}
-
-        sites_widget = QtWidgets.QWidget(self)
-        sites_layout = QtWidgets.QHBoxLayout(sites_widget)
-        active_site_combo = QtWidgets.QComboBox(sites_widget)
-        remote_site_combo = QtWidgets.QComboBox(sites_widget)
-        sites_layout.addWidget(QtWidgets.QLabel("Active Site", sites_widget))
-        sites_layout.addWidget(active_site_combo)
-        sites_layout.addWidget(QtWidgets.QLabel("Remote Site", sites_widget))
-        sites_layout.addWidget(remote_site_combo)
-
-        main_layout = QtWidgets.QVBoxLayout(self)
-        main_layout.addWidget(sites_widget)
-        main_layout.addWidget(SpacerWidget(self), 1)
-
-    def set_value(self, value):
-        pass
-        # if not value:
-        #     value = {}
-        #
-        # for root_name, widget in self.widgts_by_root_name.items():
-        #     root_value = value.get(root_name) or ""
-        #     widget.setText(root_value)
-
-    def settings_value(self):
-        return {}
-        # output = {}
-        # for root_name, widget in self.widgts_by_root_name.items():
-        #     value = widget.text()
-        #     if value:
-        #         output[root_name] = value
-        # if not output:
-        #     return None
-        # return output
-
-    def change_project(self, project_name):
-        pass
-
-
-class _ProjectListWidget(ProjectListWidget):
-    def on_item_clicked(self, new_index):
-        new_project_name = new_index.data(QtCore.Qt.DisplayRole)
-        if new_project_name is None:
-            return
-
-        if self.current_project == new_project_name:
-            return
-
-        self.select_project(new_project_name)
-        self.current_project = new_project_name
-        self.project_changed.emit()
-
-
-class ProjectSettingsWidget(QtWidgets.QWidget):
-    def __init__(self, project_settings, parent):
-        super(ProjectSettingsWidget, self).__init__(parent)
-
-        self.per_project_settings = {}
-
-        projects_widget = _ProjectListWidget(self)
-        project_specific_widget = ProjectSpecificWidget(project_settings, self)
-
-        main_layout = QtWidgets.QHBoxLayout(self)
-        main_layout.setContentsMargins(0, 0, 0, 0)
-        main_layout.addWidget(projects_widget, 0)
-        main_layout.addWidget(project_specific_widget, 1)
-
-        projects_widget.project_changed.connect(self._on_project_change)
-        # project_specific_widget.value_changed.connect(
-        #     self._on_root_value_change
-        # )
-
-        self.project_settings = project_settings
-
-        self.projects_widget = projects_widget
-        self.project_specific_widget = project_specific_widget
-
-    def _current_value(self):
-        roots_value = self.project_specific_widget.settings_value()
-        current_value = {}
-        if roots_value:
-            current_value[LOCAL_ROOTS_KEY] = roots_value
-        return current_value
-
-    def project_name(self):
-        return self.projects_widget.project_name()
-
-    def _on_project_change(self):
-        project_name = self.project_name()
-
-        self.project_settings.change_project(project_name)
-        self.project_specific_widget.change_project(project_name)
-
-    def _on_root_value_change(self):
-        self.per_project_settings[self.project_name()] = (
-            self._current_value()
-        )
-
-    def set_value(self, value):
-        if not value:
-            value = {}
-        self.per_project_settings = value
-
-        self.projects_widget.refresh()
-
-        self.project_specific_widget.set_value(self.per_project_settings)
-
-    def settings_value(self):
-        output = {}
-        for project_name, value in self.per_project_settings.items():
-            if value:
-                output[project_name] = value
-        if not output:
-            return None
         return output
